refactor(dashboard): split agents router and agent-detail view by feature
Mechanical refactor with no behavior change.

routes/agents.ts (441 lines) becomes a 22-line composition over per-action
modules under routes/agents/{new,delete,star,list,detail,tabs}.ts. Mounting
order preserved (literal /agents/new before /agents/:name).

views/agent-detail-v2.ts (466 lines) becomes a barrel re-export over
per-tab renderers under views/agent-detail/{shell,overview,nodes,config,runs,yaml}.ts.
Existing imports continue to work via the barrel.

Tests: 752/752 green. All agent routes verified 200.
