+++ packages/dashboard/src/views/agent-detail-v2.ts
@@ -1,457 +1,22 @@
 /**
- * Agent detail page — tabbed layout.
+ * Agent detail page — barrel re-export.
  *
- * Tabs:
- *   Overview  /agents/:id         — DAG + stats + recent runs + suggest
- *   Nodes     /agents/:id/nodes   — node table + add/edit/delete
- *   Config    /agents/:id/config  — LLM, variables, secrets, status
- *   Runs      /agents/:id/runs    — paginated run history
- *   YAML      /agents/:id/yaml    — already exists (separate route)
+ * The implementation lives in `agent-detail/{shell,overview,nodes,config,runs,yaml}.ts`.
+ * Routes import from here for backwards compatibility; new code can import
+ * from the per-tab modules directly.
  */
 
+export type { AgentTab, AgentDetailArgs } from './agent-detail/shell.js';
+export { agentTabStrip } from './agent-detail/shell.js';
+export { renderAgentOverview } from './agent-detail/overview.js';
+export { renderAgentNodes } from './agent-detail/nodes.js';
+export { renderAgentConfig } from './agent-detail/config.js';
+export { renderAgentRuns } from './agent-detail/runs.js';
+export { renderAgentYaml } from './agent-detail/yaml.js';
+
 import type { Agent, Run, SecretsStore } from '@some-useful-agents/core';
-import { html, render, type SafeHtml } from './html.js';
-import { layout } from './layout.js';
-import { pageHeader, type PageHeaderBack } from './page-header.js';
-import { sourceBadge, statusBadge, formatDuration, formatAge, cronToHuman } from './components.js';
-import { renderDagView, renderDagFallback } from './dag-view.js';
-import { renderOutputWidget } from './output-widgets.js';
-import {
-  renderRunInputsForm,
-  renderVariablesEditor,
-  renderOutputWidgetEditor,
-  renderNotifyEditor,
-  vStatusBadge,
-  statusOption,
-  providerOption,
-  renderModelOptions,
-  oneLine,
-} from './agent-detail-helpers.js';
-
-// ── Tab types ────────────────────────────────────────────────────────────
-
-export type AgentTab = 'overview' | 'nodes' | 'config' | 'runs' | 'yaml';
-
-export interface AgentDetailArgs {
-  agent: Agent;
-  recentRuns: Run[];
-  secretsStore: SecretsStore;
-  flash?: { kind: 'error' | 'info' | 'ok'; message: string };
-  back?: PageHeaderBack;
-  from?: string;
-  activeTab: AgentTab;
-  /** Previous run's agent-level inputs, for pre-filling the Run Now modal. */
-  previousInputs?: Record<string, string>;
-}
-
-// ── Tab strip ────────────────────────────────────────────────────────────
-
-export function agentTabStrip(agentId: string, active: AgentTab): SafeHtml {
-  const tabs: Array<{ id: AgentTab; label: string; href: string }> = [
-    { id: 'overview', label: 'Overview', href: `/agents/${agentId}` },
-    { id: 'nodes', label: 'Nodes', href: `/agents/${agentId}/nodes` },
-    { id: 'config', label: 'Config', href: `/agents/${agentId}/config` },
-    { id: 'runs', label: 'Runs', href: `/agents/${agentId}/runs` },
-    { id: 'yaml', label: 'YAML', href: `/agents/${agentId}/yaml` },
-  ];
-  return html`
-    <nav class="tab-strip">
-      ${tabs.map((t) => html`<a href="${t.href}" class="${t.id === active ? 'is-active' : ''}">${t.label}</a>`) as unknown as SafeHtml[]}
-    </nav>
-  `;
-}
-
-// ── Page shell ───────────────────────────────────────────────────────────
-
-function agentPageShell(args: AgentDetailArgs, content: SafeHtml): string {
-  const { agent, flash, back, from } = args;
-  const source = agent.source;
-  const hasCommunityShellNode = source === 'community' && agent.nodes.some((n) => n.type === 'shell');
-
-  const fromHidden = from ? html`<input type="hidden" name="from" value="${from}">` : html``;
-  const runNowButton = hasCommunityShellNode
-    ? html`<form method="POST" action="/agents/${agent.id}/run">
-        <input type="hidden" name="confirm_community_shell" value="yes">${fromHidden}
-        <button type="submit" class="btn btn--warn" onclick="return confirm('This agent contains community shell nodes. Run anyway?');">Run now (community)</button>
-      </form>`
-    : Object.keys(agent.inputs ?? {}).length > 0
-    ? html`<button type="button" class="btn btn--primary" id="run-with-inputs-btn">Run now</button>`
-    : html`<form method="POST" action="/agents/${agent.id}/run" style="display: inline;" data-run-form="${agent.id}">${fromHidden}<button type="submit" class="btn btn--primary">Run now</button></form>`;
-
-  const warningBanner = source === 'community'
-    ? html`<div class="community-banner"><strong>Community agent.</strong> Read the DAG before running.</div>`
-    : html``;
-
-  const body = html`
-    ${pageHeader({
-      title: agent.id,
-      meta: [
-        html`<form method="POST" action="/agents/${agent.id}/star" style="display:inline;margin:0;">
-          <button type="submit" class="btn-star${agent.starred ? ' is-starred' : ''}" title="${agent.starred ? 'Unstar' : 'Star'}" aria-label="${agent.starred ? 'Unstar' : 'Star'}">${agent.starred ? '\u2605' : '\u2606'}</button>
-        </form>`,
-        vStatusBadge(agent.status),
-        sourceBadge(source),
-      ],
-      cta: html`<span style="display: inline-flex; gap: var(--space-2);">
-        <button type="button" class="btn btn--ghost btn--sm" id="suggest-btn" data-agent-id="${agent.id}">Suggest improvements</button>
-        <a class="btn btn--ghost btn--sm" href="/agents/${agent.id}/versions">Versions</a>
-        ${runNowButton}
-      </span>`,
-      description: agent.description ?? undefined,
-      back,
-    })}
-    ${warningBanner}
-    ${agentTabStrip(agent.id, args.activeTab)}
-    <div class="agent-detail__tab-content">
-      ${content}
-    </div>
-
-    <div id="suggest-modal" class="modal-backdrop">
-      <div class="modal" style="max-width: 720px; max-height: 85vh; overflow-y: auto;">
-        <div id="suggest-modal-content"></div>
-      </div>
-    </div>
-
-    <div id="run-modal" class="modal-backdrop">
-      <div class="modal" style="max-width: 500px;">
-        <div id="run-modal-content">
-          ${renderRunInputsForm(agent, from, args.previousInputs)}
-        </div>
-      </div>
-    </div>
-  `;
-
-  return render(layout({ title: agent.id, activeNav: 'agents', flash }, body));
-}
-
-// ── Overview tab ─────────────────────────────────────────────────────────
-
-export async function renderAgentOverview(args: AgentDetailArgs): Promise<string> {
-  const { agent, recentRuns } = args;
-  const latestCompletedRun = recentRuns.find((r) => r.status === 'completed');
-  const hasCommunityShellNode = agent.source === 'community' && agent.nodes.some((n) => n.type === 'shell');
-
-  // Quick stats row
-  const toolIds = new Set<string>();
-  for (const n of agent.nodes) toolIds.add(n.tool ?? (n.type === 'shell' ? 'shell-exec' : 'claude-code'));
-  const toolBadges = [...toolIds].sort().map((id) => html`<a href="/tools/${id}" class="badge badge--muted" style="text-decoration: none;">${id}</a>`);
-
-  const runRows = recentRuns.slice(0, 5).map((r) => html`
-    <tr>
-      <td><a href="/runs/${r.id}" class="mono">${r.id.slice(0, 8)}</a></td>
-      <td>${statusBadge(r.status)}</td>
-      <td class="dim">${formatAge(r.startedAt)}</td>
-      <td class="dim">${formatDuration(r.startedAt, r.completedAt)}</td>
-    </tr>
-  `);
-
-  const content = html`
-    <!-- DAG + Widget preview (side-by-side when widget exists) -->
-    ${agent.outputWidget ? html`
-      <div class="run-detail-grid">
-        <div>
-          ${renderDagFallback(agent)}
-          ${renderDagView({
-            agent,
-            editBase: `/agents/${agent.id}/nodes`,
-            replay: latestCompletedRun
-              ? { priorRunId: latestCompletedRun.id, requiresCommunityConfirm: hasCommunityShellNode }
-              : undefined,
-          })}
-        </div>
-        <div class="card" style="padding: var(--space-4);">
-          <div style="display: flex; align-items: center; gap: var(--space-2); margin-bottom: var(--space-3);">
-            <h3 style="margin: 0;">Output widget</h3>
-            <span class="badge badge--muted" style="font-size: 9px;">${agent.outputWidget.type}</span>
-          </div>
-          ${latestCompletedRun?.result
-            ? html`
-              ${renderOutputWidget(agent.outputWidget, latestCompletedRun.result, agent.id) ?? html`<p class="dim" style="font-size: var(--font-size-xs);">No output to preview.</p>`}
-              <p class="dim" style="font-size: var(--font-size-xs); margin: var(--space-3) 0 0;">From run <a href="/runs/${latestCompletedRun.id}" class="mono">${latestCompletedRun.id.slice(0, 8)}</a></p>
-            `
-            : html`<p class="dim" style="font-size: var(--font-size-xs); margin: 0;">Run the agent to see a preview.</p>`}
-        </div>
-      </div>
-    ` : html`
-      ${renderDagFallback(agent)}
-      ${renderDagView({
-        agent,
-        editBase: `/agents/${agent.id}/nodes`,
-        replay: latestCompletedRun
-          ? { priorRunId: latestCompletedRun.id, requiresCommunityConfirm: hasCommunityShellNode }
-          : undefined,
-      })}
-    `}
-
-    <!-- Quick stats -->
-    <div class="agent-stats">
-      <dl class="kv">
-        <dt>Version</dt><dd class="mono">v${String(agent.version)} <a href="/agents/${agent.id}/versions" class="dim" style="font-size: var(--font-size-xs);">history</a></dd>
-        <dt>Provider</dt><dd class="mono">${agent.provider ?? 'claude'} / ${agent.model ?? 'default'}</dd>
-        <dt>Schedule</dt><dd>${agent.schedule ? html`<span title="${agent.schedule}">${cronToHuman(agent.schedule)}</span>` : html`<span class="dim">none</span>`}</dd>
-        <dt>MCP</dt><dd>${agent.mcp ? 'exposed' : html`<span class="dim">not exposed</span>`}</dd>
-        <dt>Nodes</dt><dd>${String(agent.nodes.length)}</dd>
-      </dl>
-      <div style="display: flex; gap: var(--space-2); flex-wrap: wrap; margin-top: var(--space-2);">
-        ${toolBadges as unknown as SafeHtml[]}
-      </div>
-    </div>
-
-    <!-- Recent runs -->
-    <section>
-      <h2>Recent runs</h2>
-      ${recentRuns.length === 0
-        ? html`<p class="dim">No runs yet. Hit "Run now" to get started.</p>`
-        : html`
-          <table class="table">
-            <thead><tr><th>ID</th><th>Status</th><th>Started</th><th>Duration</th></tr></thead>
-            <tbody>${runRows as unknown as SafeHtml[]}</tbody>
-          </table>
-          ${recentRuns.length > 5 ? html`<p style="margin-top: var(--space-2);"><a href="/agents/${agent.id}/runs" style="font-size: var(--font-size-xs);">View all runs &rarr;</a></p>` : html``}
-        `}
-    </section>
-
-    <!-- Danger zone — hard delete. Hidden behind a disclosure so it
-         can't be hit by an idle scroll-and-click. The form's input
-         pattern requires the operator to type the agent id verbatim
-         before the browser will submit. -->
-    <section style="margin-top: var(--space-6); border-top: 1px solid var(--color-border); padding-top: var(--space-4);">
-      <details>
-        <summary style="cursor: pointer; color: var(--color-err); font-weight: var(--weight-bold);">Danger zone</summary>
-        <div class="card" style="padding: var(--space-4); margin-top: var(--space-3); border-color: var(--color-err);">
-          <p style="margin-top: 0;">
-            <strong>Delete this agent.</strong>
-            All ${String(agent.nodes.length)}-node ${agent.source === 'community' ? 'community ' : ''}DAG and every prior version are removed.
-            <span class="dim">Run history is kept (orphaned — runs reference the agent by id, no FK).</span>
-          </p>
-          <form method="POST" action="/agents/${agent.id}/delete" style="display: flex; gap: var(--space-2); align-items: center; flex-wrap: wrap;">
-            <label style="font-size: var(--font-size-xs);">
-              Type <code>${agent.id}</code> to confirm:
-              <input type="text" name="confirm" required
-                pattern="^${agent.id.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')}$"
-                placeholder="${agent.id}"
-                style="margin-left: var(--space-2); font-family: var(--font-mono);" />
-            </label>
-            <button type="submit" class="btn btn--sm" style="color: var(--color-err); border-color: var(--color-err);">
-              Delete forever
-            </button>
-          </form>
-        </div>
-      </details>
-    </section>
-  `;
-
-  return agentPageShell({ ...args, activeTab: 'overview' }, content);
-}
-
-// ── Nodes tab ────────────────────────────────────────────────────────────
-
-export async function renderAgentNodes(args: AgentDetailArgs): Promise<string> {
-  const { agent, secretsStore } = args;
-
-  const nodeRows = agent.nodes.map((n) => {
-    const body = n.type === 'shell' ? oneLine(n.command ?? '') : oneLine(n.prompt ?? '');
-    const deps = n.dependsOn?.length ? n.dependsOn.join(', ') : html`<span class="dim">\u2014</span>`;
-    const secrets = n.secrets?.length ? n.secrets.join(', ') : html`<span class="dim">\u2014</span>`;
-    return html`
-      <tr id="node-${n.id}">
-        <td class="mono">${n.id}</td>
-        <td>${n.type === 'shell' ? html`<span class="badge badge--ok">shell</span>` : html`<span class="badge badge--info">claude-code</span>`}</td>
-        <td class="mono">${deps}</td>
-        <td class="mono">${secrets}</td>
-        <td class="dim">${body}</td>
-        <td style="white-space: nowrap;">
-          <a class="btn btn--sm" href="/agents/${agent.id}/nodes/${n.id}/edit">Edit</a>
-          <form method="POST" action="/agents/${agent.id}/nodes/${n.id}/delete" style="display: inline; margin: 0;"
-                data-confirm="Delete node '${n.id}'?">
-            <button type="submit" class="btn btn--sm btn--ghost" style="color: var(--color-err);">Delete</button>
-          </form>
-        </td>
-      </tr>
-    `;
-  });
-
-  // Secrets by node
-  const secretRows: SafeHtml[] = [];
-  for (const node of agent.nodes) {
-    for (const name of node.secrets ?? []) {
-      let present: 'ok' | 'missing' | 'unknown';
-      try { present = (await secretsStore.has(name)) ? 'ok' : 'missing'; } catch { present = 'unknown'; }
-      const badge = present === 'ok' ? html`<span class="badge badge--ok">set</span>`
-        : present === 'missing' ? html`<span class="badge badge--err">missing</span>`
-        : html`<span class="badge badge--warn">locked</span>`;
-      secretRows.push(html`<tr><td class="mono">${node.id}</td><td class="mono">${name}</td><td>${badge}</td></tr>`);
-    }
-  }
-
-  const content = html`
-    <div style="display: flex; justify-content: space-between; align-items: center; margin-bottom: var(--space-4);">
-      <h2 style="margin: 0;">Nodes</h2>
-      <a class="btn btn--primary btn--sm" href="/agents/${agent.id}/add-node">+ Add node</a>
-    </div>
-    <table class="table">
-      <thead><tr><th>Id</th><th>Type</th><th>Depends on</th><th>Secrets</th><th>Body</th><th></th></tr></thead>
-      <tbody>${nodeRows as unknown as SafeHtml[]}</tbody>
-    </table>
-
-    ${secretRows.length > 0 ? html`
-      <section style="margin-top: var(--space-6);">
-        <h2>Secrets by node</h2>
-        <table class="table">
-          <thead><tr><th>Node</th><th>Secret</th><th>Status</th></tr></thead>
-          <tbody>${secretRows}</tbody>
-        </table>
-      </section>
-    ` : html``}
-  `;
-
-  return agentPageShell({ ...args, activeTab: 'nodes' }, content);
-}
-
-// ── Config tab ───────────────────────────────────────────────────────────
-
-export async function renderAgentConfig(args: AgentDetailArgs): Promise<string> {
-  const { agent, secretsStore } = args;
-
-  // Secret counts for the secrets section
-  let secretsSet = 0;
-  let secretsMissing = 0;
-  const allSecrets = new Set<string>();
-  for (const node of agent.nodes) {
-    for (const name of node.secrets ?? []) allSecrets.add(name);
-  }
-  for (const name of allSecrets) {
-    try { if (await secretsStore.has(name)) secretsSet++; else secretsMissing++; } catch { /* unknown */ }
-  }
-
-  const content = html`
-    <!-- Status -->
-    <section class="card" style="margin-bottom: var(--space-6);">
-      <h3 style="margin: 0 0 var(--space-3);">Status</h3>
-      <form method="POST" action="/agents/${agent.id}/status" style="display: flex; gap: var(--space-2); align-items: center;">
-        <select name="newStatus" class="form-field" style="padding: var(--space-1) var(--space-2); font-size: var(--font-size-sm);">
-          ${statusOption('active', agent.status)}
-          ${statusOption('paused', agent.status)}
-          ${statusOption('draft', agent.status)}
-          ${statusOption('archived', agent.status)}
-        </select>
-        <button type="submit" class="btn btn--sm">Apply</button>
-      </form>
-    </section>
-
-    <!-- LLM defaults -->
-    <section class="card" style="margin-bottom: var(--space-6);">
-      <h3 style="margin: 0 0 var(--space-3);">LLM defaults</h3>
-      <form method="POST" action="/agents/${agent.id}/llm" id="llm-form" style="display: flex; flex-direction: column; gap: var(--space-2);">
-        <div style="display: flex; gap: var(--space-2); align-items: center;">
-          <label style="font-size: var(--font-size-xs); color: var(--color-text-muted); min-width: 55px;">Provider</label>
-          <select name="provider" id="llm-provider" class="form-field" style="flex: 1; padding: var(--space-1) var(--space-2); font-size: var(--font-size-sm);">
-            ${providerOption('claude', agent.provider)}
-            ${providerOption('codex', agent.provider)}
-          </select>
-        </div>
-        <div style="display: flex; gap: var(--space-2); align-items: center;">
-          <label style="font-size: var(--font-size-xs); color: var(--color-text-muted); min-width: 55px;">Model</label>
-          <select name="model" id="llm-model" class="form-field" style="flex: 1; padding: var(--space-1) var(--space-2); font-size: var(--font-size-sm); font-family: var(--font-mono);">
-            ${renderModelOptions(agent.provider, agent.model)}
-          </select>
-        </div>
-        <div id="llm-model-desc" class="dim" style="font-size: var(--font-size-xs); min-height: 1.2em;"></div>
-        <div style="display: flex; justify-content: flex-end;">
-          <button type="submit" class="btn btn--sm">Apply</button>
-        </div>
-        <p class="dim" style="font-size: var(--font-size-xs); margin: 0;">Applies to all claude-code nodes. Individual nodes can override in YAML.</p>
-      </form>
-    </section>
-
-    <!-- Variables -->
-    <section class="card" style="margin-bottom: var(--space-6);">
-      <h3 style="margin: 0 0 var(--space-3);">Variables</h3>
-      ${renderVariablesEditor(agent)}
-    </section>
-
-    <!-- Output Widget -->
-    <section class="card" style="margin-bottom: var(--space-6);">
-      <h3 style="margin: 0 0 var(--space-3);">Output Widget</h3>
-      ${renderOutputWidgetEditor(agent)}
-    </section>
-
-    <!-- Notify -->
-    <section class="card" style="margin-bottom: var(--space-6);">
-      <h3 style="margin: 0 0 var(--space-3);">Notify</h3>
-      ${renderNotifyEditor(agent)}
-    </section>
-
-    <!-- Secrets -->
-    <section class="card">
-      <h3 style="margin: 0 0 var(--space-3);">Secrets</h3>
-      <p class="dim" style="font-size: var(--font-size-xs); margin: 0 0 var(--space-2);">
-        ${String(allSecrets.size)} declared. ${String(secretsSet)} set, ${String(secretsMissing)} missing.
-      </p>
-      <a href="/settings/secrets" class="btn btn--sm">Manage secrets</a>
-    </section>
-  `;
-
-  return agentPageShell({ ...args, activeTab: 'config' }, content);
-}
-
-// ── Runs tab ─────────────────────────────────────────────────────────────
-
-export function renderAgentRuns(args: AgentDetailArgs): string {
-  const { agent, recentRuns } = args;
-
-  const runRows = recentRuns.map((r) => html`
-    <tr>
-      <td><a href="/runs/${r.id}" class="mono">${r.id.slice(0, 8)}</a></td>
-      <td>${statusBadge(r.status)}</td>
-      <td class="dim">${formatAge(r.startedAt)}</td>
-      <td class="dim">${formatDuration(r.startedAt, r.completedAt)}</td>
-      <td class="dim">${r.triggeredBy}</td>
-    </tr>
-  `);
-
-  const content = html`
-    <h2>Run history</h2>
-    ${recentRuns.length === 0
-      ? html`<p class="dim">No runs yet.</p>`
-      : html`
-        <table class="table">
-          <thead><tr><th>ID</th><th>Status</th><th>Started</th><th>Duration</th><th>Triggered</th></tr></thead>
-          <tbody>${runRows as unknown as SafeHtml[]}</tbody>
-        </table>
-      `}
-  `;
-
-  return agentPageShell({ ...args, activeTab: 'runs' }, content);
-}
-
-// ── YAML tab ─────────────────────────────────────────────────────────────
-
-export function renderAgentYaml(args: AgentDetailArgs & { yaml: string; error?: string }): string {
-  const { yaml, error } = args;
-  const agent = args.agent;
-
-  const content = html`
-    ${error ? html`<div class="flash flash--error">${error}</div>` : html``}
-    <p class="dim" style="font-size: var(--font-size-xs); margin: 0 0 var(--space-3);">
-      v${String(agent.version)}. Saving creates a new version. The YAML is validated before save.
-    </p>
-    <form method="POST" action="/agents/${agent.id}/yaml" class="card" style="max-width: 800px;">
-      <label style="display: flex; flex-direction: column; gap: var(--space-2);">
-        <textarea name="yaml" rows="30" required
-          style="padding: var(--space-3); border: 1px solid var(--color-border-strong); border-radius: var(--radius-sm); font-family: var(--font-mono); font-size: var(--font-size-xs); resize: vertical; line-height: 1.5; tab-size: 2;">${yaml}</textarea>
-      </label>
-      <div style="margin-top: var(--space-3); display: flex; gap: var(--space-2); justify-content: flex-end;">
-        <a class="btn btn--ghost" href="/agents/${agent.id}">Cancel</a>
-        <button type="submit" class="btn btn--primary">Save YAML</button>
-      </div>
-    </form>
-  `;
-
-  return agentPageShell({ ...args, activeTab: 'yaml' }, content);
-}
-
-// ── Legacy export (backward compat for the route) ────────────────────────
+import type { PageHeaderBack } from './page-header.js';
+import { renderAgentOverview } from './agent-detail/overview.js';
 
 /** @deprecated Use renderAgentOverview directly */
 export async function renderAgentDetailV2(args: {
